arrayspc {elasticnet}	R Documentation

Sparse PCs of Microarrays

Description

Sparse PC by iterative SVD and soft-thresholding

Usage

arrayspc(x,K=1,para,use.corr=FALSE, max.iter=100,trace=FALSE,eps=1e-3)

Arguments

x
The microarray matrix.

K
Number of components. Default is 1.

para
The thresholding parameters. A vector of length K.

use.corr
Perform PCA on the correlation matrix? This option is only effective when the argument type is set "data".

max.iter
Maximum number of iterations.

trace
If TRUE, prints out its progress.

eps
Convergence criterion.

Details

The function is equivalent to a special case of spca() with the quadratic penalty=infinity. It is specifically designed for the case p>>n, like microarrays.

Value

A "arrayspc" object is returned.

Author(s)

Hui Zou, Trevor Hastie and Robert Tibshirani

References

Zou, H., Hastie, T. and Tibshirani, R. (2004) "Sparse principal component analysis" Technical report, Statistics Dept. Stanford University

See Also

spca, princomp

[Package elasticnet version 1.1 Index]