pitprops {elasticnet}	R Documentation

Pitprops correlation data

Description

The pitprops data is a correlation matrix that was calculated from 180 observations. There are 13 explanatory variables.

Usage

data(pitprops)

Details

Jeffers (1967) tried to interpret the first six PCs. This is a classical example showing the difficulty of interpreting principal components.

References

Jeffers, J. (1967) "Two case studies in the application of principal component", Applied Statistics, 16, 225-236.

[Package elasticnet version 1.1 Index]